+++ core-ng/src/test/resources/db-test/update-query-auto-increment-id.java
@@ -1,5 +1,5 @@
 public class UpdateQuery$AutoIncrementIdEntity implements core.framework.internal.db.UpdateQuery {
-    public core.framework.internal.db.UpdateQuery.Statement update(Object value, boolean partial) {
+    public core.framework.internal.db.UpdateQuery.Statement update(Object value, boolean partial, String where, Object[] whereParams) {
         core.framework.internal.db.AutoIncrementIdEntity entity = (core.framework.internal.db.AutoIncrementIdEntity) value;
         if (entity.id == null) throw new Error("primary key must not be null, field=id");
         StringBuilder sql = new StringBuilder("UPDATE auto_increment_id_entity SET ");
@@ -37,6 +37,10 @@ public class UpdateQuery$AutoIncrementIdEntity implements core.framework.interna
         }
         sql.append(" WHERE id = ?");
         params.add(entity.id);
+        if (where != null) {
+            sql.append(" AND (").append(where).append(')');
+            for (int i = 0; i< whereParams.length; i++) params.add(whereParams[i]);
+        }
         return new core.framework.internal.db.UpdateQuery.Statement(sql.toString(), params.toArray());
     }